requirements.txt
avro==1.11.3
cassandra-driver==3.29.0
click==8.1.7
confluent-kafka==2.3.0
cqlengine==0.21.0
fastavro==1.9.4
geomet==0.2.1.post1
six==1.16.0
